+++ model_fetch/lib/src/model_by_id/fixed_model_by_id_bloc.dart
@@ -38,6 +38,9 @@ class FixedModelByIdBloc<
   }
 
   @override
+  Future<T> get() async => model;
+
+  @override
   void reload() {
     model = _getter();
     emitState(