+++ model_fetch_firestore/lib/src/model_by_filter/firestore_live_by_filter_bloc.dart
@@ -1,20 +1,29 @@
+import 'dart:async';
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:collection/collection.dart';
 import 'package:model_fetch/model_fetch.dart';
 import 'package:model_interfaces/model_interfaces.dart';
 
+import '../util.dart';
+
 class FirestoreLiveByFilterBloc<T extends WithId<String>>
     extends ModelByFilterBloc<String, T> {
-  final Query<T> query;
+  final ErrorCallback onError;
+  final Query<Future<T>> query;
+
+  StreamSubscription? _subscription;
 
   FirestoreLiveByFilterBloc({
+    required this.onError,
     required this.query,
   }) {
-    query.snapshots().listen(_onModelChanged);
+    _subscription =
+        query.snapshots().handleError(onError).listen(_onModelChanged);
   }
 
-  void _onModelChanged(QuerySnapshot<T?> querySnapshot) {
-    final model = querySnapshot.docs.firstOrNull?.data();
+  Future<void> _onModelChanged(QuerySnapshot<Future<T?>> querySnapshot) async {
+    final model = await querySnapshot.docs.firstOrNull?.data();
 
     emitStateIfChanged(
       ModelByFilterState(
@@ -22,5 +31,11 @@ class FirestoreLiveByFilterBloc<T extends WithId<String>>
         status: model == null ? LoadStatus.error : LoadStatus.ok,
       ),
     );
+  }
+
+  @override
+  Future<void> dispose() async {
+    await _subscription?.cancel();
+    await super.dispose();
   }
 }