+++ model_fetch_firestore/lib/src/util.dart
@@ -1,0 +1,1 @@
+typedef ErrorCallback = void Function(Object error, StackTrace trace);